+++ src/components/button/button_content.tsx
@@ -58,6 +58,7 @@ export interface EuiButtonContentProps extends CommonProps {
       ref?: Ref<HTMLSpanElement>;
       'data-text'?: string;
     };
+  iconSize?: 's' | 'm';
 }
 
 export const EuiButtonContent: FunctionComponent<
@@ -67,6 +68,7 @@ export const EuiButtonContent: FunctionComponent<
   textProps,
   isLoading = false,
   iconType,
+  iconSize = 'm',
   iconSide = 'left',
   ...contentProps
 }) => {
@@ -82,7 +84,7 @@ export const EuiButtonContent: FunctionComponent<
       <EuiIcon
         className="euiButtonContent__icon"
         type={iconType}
-        size="m"
+        size={iconSize}
         color="default" // forces two-tone icons to inherit their parent color
       />
     );